Rename `inbox.basicauth` to `inbox.exceptions`.
The comment at the top of this module already said:
> TODO(emfree): this is now legitimately just a grab-bag of nebulous
> exceptions. Rename module and clean up.
