+++ inbox/models/backends/oauth.py
@@ -10,7 +10,7 @@ from sqlalchemy import Column, ForeignKey
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.orm import relationship
 
-from inbox.basicauth import OAuthError
+from inbox.exceptions import OAuthError
 from inbox.logging import get_logger
 from inbox.models.secret import Secret, SecretType
