+++ inbox/search/backends/gmail.py
@@ -5,7 +5,7 @@ from sqlalchemy import desc
 
 from inbox.api.kellogs import APIEncoder
 from inbox.auth.oauth import OAuthRequestsWrapper
-from inbox.basicauth import OAuthError
+from inbox.exceptions import OAuthError
 from inbox.logging import get_logger
 from inbox.models import Account, Message, Thread
 from inbox.models.backends.oauth import token_manager
