+++ tests/auth/providers/mock_gmail.py
@@ -12,7 +12,7 @@ foobar+no_all_mail@gmail.com.
 """
 
 from inbox.auth.gmail import GmailAuthHandler
-from inbox.basicauth import (
+from inbox.exceptions import (
     GmailSettingError,
     ImapSupportDisabledError,
     OAuthError,
