+++ tests/imap/test_full_imap_enabled.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 from imapclient import IMAPClient
 
 from inbox.auth.generic import GenericAuthHandler
-from inbox.basicauth import UserRecoverableConfigError
+from inbox.exceptions import UserRecoverableConfigError
 
 
 class MockIMAPClient(IMAPClient):
